[FRONTEND] `dot` now uses `tl.float32` by default for `out_dtype`.

+++ python/triton/language/core.py
@@ -862,7 +862,7 @@ def reshape(input, shape, _builder=None):
 
 
 @builtin
-def dot(input, other, allow_tf32=True, out_dtype=float16, _builder=None):
+def dot(input, other, allow_tf32=True, out_dtype=float32, _builder=None):
     """
     Returns the matrix product of two blocks.
